+++ com.unity.render-pipelines.high-definition/Editor/Lighting/VolumetricLighting/VolumetricCloudsEditor.cs
@@ -17,6 +17,9 @@ namespace UnityEditor.Rendering.HighDefinition
         SerializedDataParameter m_CloudControl;
 
         SerializedDataParameter m_CloudPreset;
+        SerializedDataParameter m_CustomDensityCurve;
+        SerializedDataParameter m_CustomErosionCurve;
+        SerializedDataParameter m_CustomAmbientOcclusionCurve;
 
         SerializedDataParameter m_CumulusMap;
         SerializedDataParameter m_CumulusMapMultiplier;
@@ -48,6 +51,7 @@ namespace UnityEditor.Rendering.HighDefinition
         SerializedDataParameter m_PowderEffectIntensity;
         SerializedDataParameter m_MultiScattering;
         SerializedDataParameter m_AmbientLightProbeDimmer;
+        SerializedDataParameter m_ErosionOcclusion;
 
         // Wind
         SerializedDataParameter m_GlobalWindSpeed;
@@ -81,6 +85,9 @@ namespace UnityEditor.Rendering.HighDefinition
             m_CloudControl = Unpack(o.Find(x => x.cloudControl));
 
             m_CloudPreset = Unpack(o.Find(x => x.cloudPreset));
+            m_CustomDensityCurve = Unpack(o.Find(x => x.customDensityCurve));
+            m_CustomErosionCurve = Unpack(o.Find(x => x.customErosionCurve));
+            m_CustomAmbientOcclusionCurve = Unpack(o.Find(x => x.customAmbientOcclusionCurve));
 
             m_CumulusMap = Unpack(o.Find(x => x.cumulusMap));
             m_CumulusMapMultiplier = Unpack(o.Find(x => x.cumulusMapMultiplier));
@@ -112,7 +119,8 @@ namespace UnityEditor.Rendering.HighDefinition
             m_PowderEffectIntensity = Unpack(o.Find(x => x.powderEffectIntensity));
             m_MultiScattering = Unpack(o.Find(x => x.multiScattering));
             m_AmbientLightProbeDimmer = Unpack(o.Find(x => x.ambientLightProbeDimmer));
-
+            m_ErosionOcclusion = Unpack(o.Find(x => x.erosionOcclusion));
+            
             // Wind
             m_GlobalWindSpeed = Unpack(o.Find(x => x.globalWindSpeed));
             m_Orientation = Unpack(o.Find(x => x.orientation));
@@ -192,6 +200,12 @@ namespace UnityEditor.Rendering.HighDefinition
                 {
                     using (new HDEditorUtils.IndentScope(needsIntendation ? 16 : 0))
                     {
+                        if (controlMode == VolumetricClouds.CloudControl.Simple)
+                        {
+                            PropertyField(m_CustomDensityCurve);
+                            PropertyField(m_CustomErosionCurve);
+                            PropertyField(m_CustomAmbientOcclusionCurve);
+                        }
                         PropertyField(m_DensityMultiplier);
                         PropertyField(m_ShapeFactor);
                         PropertyField(m_ShapeScale);
@@ -226,6 +240,7 @@ namespace UnityEditor.Rendering.HighDefinition
             DrawHeader("Lighting");
             {
                 PropertyField(m_AmbientLightProbeDimmer);
+                PropertyField(m_ErosionOcclusion);
                 PropertyField(m_ScatteringTint);
                 PropertyField(m_PowderEffectIntensity);
                 PropertyField(m_MultiScattering);