+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/VolumetricLighting/HDRenderPipeline.VolumetricClouds.cs
@@ -15,6 +15,9 @@ namespace UnityEngine.Rendering.HighDefinition
         Texture2D m_CloudyPresetMap;
         Texture2D m_OvercastPresetMap;
         Texture2D m_StormyPresetMap;
+        Texture2D m_CustomLutPresetMap;
+        const int k_CustomLutMapResolution = 32;
+        readonly Color[] m_CustomLutColorArray = new Color[k_CustomLutMapResolution];
 
         // The set of kernels that are required
         int m_ConvertObliqueDepthKernel;
@@ -79,6 +82,48 @@ namespace UnityEngine.Rendering.HighDefinition
             m_StormyPresetMap = new Texture2D(1, 1, GraphicsFormat.R8G8B8A8_UNorm, TextureCreationFlags.None) { name = "Default Storm Texture" };
             m_StormyPresetMap.SetPixel(0, 0, new Color(1.0f, 0.0f, 0.375f, 1.0f));
             m_StormyPresetMap.Apply();
+        }
+
+        void PrepareCustomLutData(in VolumetricClouds clouds)
+        {
+
+            if (m_CustomLutPresetMap == null)
+            {
+                m_CustomLutPresetMap = new Texture2D(1, k_CustomLutMapResolution, GraphicsFormat.R16G16B16A16_SFloat, TextureCreationFlags.None)
+                {
+                    name = "Custom LUT Curve",
+                    filterMode = FilterMode.Bilinear,
+                    wrapMode = TextureWrapMode.Clamp
+                };
+                m_CustomLutPresetMap.hideFlags = HideFlags.HideAndDontSave;
+            }
+
+            var pixels = m_CustomLutColorArray;
+
+            var densityCurve = clouds.customDensityCurve.value;
+            var erosionCurve = clouds.customErosionCurve.value;
+            var ambientOcclusionCurve = clouds.customAmbientOcclusionCurve.value;
+            if (densityCurve == null || densityCurve.length == 0)
+            {
+                for (int i = 0; i < k_CustomLutMapResolution; i++)
+                    pixels[i] = Color.white;
+            }
+            else
+            {
+                float step = 1.0f / (k_CustomLutMapResolution - 1f);
+
+                for (int i = 0; i < k_CustomLutMapResolution; i++)
+                {
+                    float currTime = step * i;
+                    float density = Mathf.Clamp(densityCurve.Evaluate(currTime), 0.0f, 1.0f);
+                    float erosion = Mathf.Clamp(erosionCurve.Evaluate(currTime), 0.0f, 1.0f); ;
+                    float ambientOcclusion = Mathf.Clamp(ambientOcclusionCurve.Evaluate(currTime), 0.0f, 1.0f);
+                    pixels[i] = new Color(density, erosion, ambientOcclusion, 1.0f);
+                }
+            }
+
+            m_CustomLutPresetMap.SetPixels(pixels);
+            m_CustomLutPresetMap.Apply();
         }
 
         // Function that fills the buffer with the ambient probe values
@@ -201,58 +246,67 @@ namespace UnityEngine.Rendering.HighDefinition
             return Mathf.Sqrt((k_EarthRadius + lowerCloudRadius) * (k_EarthRadius + lowerCloudRadius) - k_EarthRadius * earthRadius);
         }
 
-        void GetPresetCloudMapValues(VolumetricClouds.CloudPresets preset, out float densityMultiplier, out float shapeFactor, out float shapeScale, out float erosionFactor, out float erosionScale)
+        void GetPresetCloudMapValues(VolumetricClouds.CloudPresets preset, out CloudModelData cloudModelData)
         {
             switch (preset)
             {
                 case VolumetricClouds.CloudPresets.Sparse:
                 {
-                    densityMultiplier = 0.8f;
-                    shapeFactor = 0.9f;
-                    shapeScale = 0.35f;
-                    erosionFactor = 0.7f;
-                    erosionScale = 0.6f;
+                    cloudModelData.densityMultiplier = 0.4472135955f;
+                    cloudModelData.shapeFactor = 0.9f;
+                    cloudModelData.shapeScale = 1.75f;
+                    cloudModelData.erosionFactor = 0.7f;
+                    cloudModelData.erosionScale = 30.0f;
                     return;
                 }
                 case VolumetricClouds.CloudPresets.Cloudy:
                 {
-                    densityMultiplier = 0.7f;
-                    shapeFactor = 0.75f;
-                    shapeScale = 0.5f;
-                    erosionFactor = 0.8f;
-                    erosionScale = 0.5f;
+                    cloudModelData.densityMultiplier = 0.41833001326f;
+                    cloudModelData.shapeFactor = 0.75f;
+                    cloudModelData.shapeScale = 2.5f;
+                    cloudModelData.erosionFactor = 0.8f;
+                    cloudModelData.erosionScale = 25.0f;
                     return;
                 }
                 case VolumetricClouds.CloudPresets.Overcast:
                 {
-                    densityMultiplier = 0.1f;
-                    shapeFactor = 0.3f;
-                    shapeScale = 0.7f;
-                    erosionFactor = 0.0f;
-                    erosionScale = 1.0f;
+                    cloudModelData.densityMultiplier = 0.158113883f;
+                    cloudModelData.shapeFactor = 0.3f;
+                    cloudModelData.shapeScale = 3.5f;
+                    cloudModelData.erosionFactor = 0.0f;
+                    cloudModelData.erosionScale = 50.0f;
                     return;
                 }
                 case VolumetricClouds.CloudPresets.Stormy:
                 {
-                    densityMultiplier = 1.25f;
-                    shapeFactor = 0.7f;
-                    shapeScale = 0.7f;
-                    erosionFactor = 0.6f;
-                    erosionScale = 1.2f;
+                    cloudModelData.densityMultiplier = 0.55901699437f;
+                    cloudModelData.shapeFactor = 0.7f;
+                    cloudModelData.shapeScale = 3.5f;
+                    cloudModelData.erosionFactor = 0.6f;
+                    cloudModelData.erosionScale = 60.0f;
                     return;
                 }
             }
 
             // Default unused values
-            densityMultiplier = 0.6f;
-            shapeFactor = 0.6f;
-            shapeScale = 1.0f;
-            erosionFactor = 0.6f;
-            erosionScale = 1.0f;
+            cloudModelData.densityMultiplier = 0.38729833462f;
+            cloudModelData.shapeFactor = 0.6f;
+            cloudModelData.shapeScale = 0.33333333333f;
+            cloudModelData.erosionFactor = 0.6f;
+            cloudModelData.erosionScale = 0.33333333333f;
         }
 
         // The earthRadius
         const float k_EarthRadius = 6378100.0f;
+
+        internal struct CloudModelData
+        {
+            public float densityMultiplier;
+            public float shapeFactor;
+            public float shapeScale;
+            public float erosionFactor;
+            public float erosionScale;
+        }
 
         void UpdateShaderVariableslClouds(ref ShaderVariablesClouds cb, HDCamera hdCamera, VolumetricClouds settings, in VolumetricCloudsParameters parameters, bool shadowPass)
         {
@@ -264,8 +318,8 @@ namespace UnityEngine.Rendering.HighDefinition
 
             cb._NumPrimarySteps = settings.numPrimarySteps.value;
             cb._NumLightSteps = settings.numLightSteps.value;
-            // 1500.0f is the maximal distance that a single step can do in theory (otherwise we endup skipping large clouds)
-            cb._MaxRayMarchingDistance = Mathf.Min(1500.0f * cb._NumPrimarySteps, hdCamera.camera.farClipPlane);
+            // 1000.0f is the maximal distance that a single step can do in theory (otherwise we endup skipping large clouds)
+            cb._MaxRayMarchingDistance = Mathf.Min(1000.0f * cb._NumPrimarySteps, hdCamera.camera.farClipPlane);
             cb._CloudMapTiling.Set(settings.cloudTiling.value.x, settings.cloudTiling.value.y, settings.cloudOffset.value.x, settings.cloudOffset.value.y);
 
             cb._ScatteringTint = Color.white - settings.scatteringTint.value * 0.75f;
@@ -319,20 +373,26 @@ namespace UnityEngine.Rendering.HighDefinition
 
             cb._MultiScattering = 1.0f - settings.multiScattering.value * 0.8f;
 
+            CloudModelData cloudModelData;
             if (settings.cloudControl.value == VolumetricClouds.CloudControl.Simple && settings.cloudPreset.value != VolumetricClouds.CloudPresets.Custom)
             {
-                GetPresetCloudMapValues(settings.cloudPreset.value, out cb._DensityMultiplier, out cb._ShapeFactor, out cb._ShapeScale, out cb._ErosionFactor, out cb._ErosionScale);
+                GetPresetCloudMapValues(settings.cloudPreset.value, out cloudModelData);
             }
             else
             {
-                // The density multiplier is not used linearly
-                float densityMultiplier = settings.densityMultiplier.value * 2.0f;
-                cb._DensityMultiplier = densityMultiplier * densityMultiplier;
-                cb._ShapeFactor = settings.shapeFactor.value;
-                cb._ShapeScale = Mathf.Lerp(0.5f, 2.0f, settings.shapeScale.value);
-                cb._ErosionFactor = settings.erosionFactor.value;
-                cb._ErosionScale = Mathf.Lerp(0.5f, 2.0f, settings.erosionScale.value);
+                cloudModelData.densityMultiplier = settings.densityMultiplier.value;
+                cloudModelData.shapeFactor = settings.shapeFactor.value;
+                cloudModelData.shapeScale = settings.shapeScale.value;
+                cloudModelData.erosionFactor = settings.erosionFactor.value;
+                cloudModelData.erosionScale = settings.erosionScale.value;
             }
+
+            // The density multiplier is not used linearly
+            cb._DensityMultiplier = cloudModelData.densityMultiplier * cloudModelData.densityMultiplier * 4.0f;
+            cb._ShapeFactor = cloudModelData.shapeFactor;
+            cb._ShapeScale = cloudModelData.shapeScale;
+            cb._ErosionFactor = cloudModelData.erosionFactor;
+            cb._ErosionScale = cloudModelData.erosionScale;
 
             // If the sun has moved more than 2.0°, reduce significantly the history accumulation
             float sunAngleDifference = 0.0f;
@@ -347,6 +407,7 @@ namespace UnityEngine.Rendering.HighDefinition
 
             float absoluteCloudHighest = cb._HighestCloudAltitude + cb._EarthRadius;
             cb._MaxCloudDistance = Mathf.Sqrt(absoluteCloudHighest * absoluteCloudHighest - cb._EarthRadius * cb._EarthRadius);
+            cb._ErosionOcclusion = settings.erosionOcclusion.value;
 
             // If this is a planar reflection, we need to compute the non oblique matrices
             if (hdCamera.camera.cameraType == CameraType.Reflection)
@@ -461,7 +522,13 @@ namespace UnityEngine.Rendering.HighDefinition
             if (settings.cloudControl.value == VolumetricClouds.CloudControl.Simple)
             {
                 parameters.cloudMapTexture = GetPresetCloudMapTexture(settings.cloudPreset.value);
-                parameters.cloudLutTexture = m_Asset.renderPipelineResources.textures.cloudLutRainAO;
+                if (settings.cloudPreset.value == VolumetricClouds.CloudPresets.Custom)
+                {
+                    PrepareCustomLutData(settings);
+                    parameters.cloudLutTexture = m_CustomLutPresetMap;
+                }
+                else
+                    parameters.cloudLutTexture = m_Asset.renderPipelineResources.textures.cloudLutRainAO;
             }
             else if (settings.cloudControl.value == VolumetricClouds.CloudControl.Advanced)
             {